simulation/tests/test_customer_generator.py
# simulation/tests/test_customer_generator.py
import pytest
import tempfile
from pathlib import Path

from simulation.customer_generator import CustomerGenerator, GeneratedCustomer  # Absolute import for tests
from simulation.profile_templates import CUSTOMER_TEMPLATES  # Absolute import for tests



class TestCustomerGenerator:
    
    def setup_method(self):
        """Setup for each test method"""
        self.generator = CustomerGenerator()
    
    def test_generate_customer_from_template(self):
        """Test generating customer from predefined template"""
        customer = self.generator.generate_customer('TEST1', 'pharmaceutical')
        
        assert customer.id == 'TEST1'
        assert customer.data_sharing_method == 'api'
        assert len(customer.facilities) >= 1
        assert len(customer.facilities) <= 2
        assert 'pharma' in customer.name.lower()
    
    def test_generate_customer_random_template(self):
        """Test generating customer with random template"""
        customer = self.generator.generate_customer('TEST2')
        
        assert customer.id == 'TEST2'
        assert customer.data_sharing_method in ['csv', 'api']
        assert len(customer.facilities) >= 1
        assert customer.name is not None
    
    def test_generate_multiple_customers(self):
        """Test generating multiple customers with distribution"""
        distribution = {
            'pharmaceutical': 0.5,
            'food_storage': 0.5
        }
        
        customers = self.generator.generate_multiple_customers(10, distribution)
        
        assert len(customers) == 10
        assert all(c.id != '' for c in customers)
        
        # Check distribution roughly matches
        pharma_count = sum(1 for c in customers if 'pharma' in c.name.lower())
        food_count = sum(1 for c in customers if 'food' in c.name.lower())
        
        assert pharma_count > 0
        assert food_count > 0
    
    def test_customer_facilities_and_units(self):
        """Test that generated customers have proper facilities and units"""
        customer = self.generator.generate_customer('TEST3', 'industrial')
        
        assert len(customer.facilities) >= 2
        
        for facility in customer.facilities:
            assert facility.customer_id == customer.id
            assert facility.country is not None
            assert len(facility.units) >= 3
            
            for unit in facility.units:
                assert unit.facility_id == facility.id
                assert unit.size > 0
                assert unit.size_unit in ['sqm', 'sqft']
                assert unit.temperature_unit in ['C', 'F']
                assert unit.data_frequency > 0
    
    def test_data_quality_settings(self):
        """Test that data quality settings are properly applied"""
        customer = self.generator.generate_customer('TEST4', 'small_business')
        
        # Small business should have lower quality
        for facility in customer.facilities:
            for unit in facility.units:
                assert unit.data_quality['equipment_reliability'] == 'low'
                assert unit.data_quality['null_reading_probability'] > 0.005
    
    def test_assignment_customers(self):
        """Test generating exact assignment customers A and B"""
        customer_a = self.generator.generate_assignment_customer('A')
        customer_b = self.generator.generate_assignment_customer('B')
        
        # Customer A specifications
        assert customer_a.id == 'A'
        assert customer_a.name == 'Customer A'
        assert customer_a.data_sharing_method == 'csv'
        assert len(customer_a.facilities) == 1
        assert len(customer_a.facilities[0].units) == 1
        
        unit_a = customer_a.facilities[0].units[0]
        assert unit_a.name is None
        assert unit_a.size == 930
        assert unit_a.size_unit == 'sqm'
        assert unit_a.set_temperature == -20
        assert unit_a.temperature_unit == 'C'
        
        # Customer B specifications
        assert customer_b.id == 'B'
        assert customer_b.name == 'Customer B'
        assert customer_b.data_sharing_method == 'api'
        assert len(customer_b.facilities) == 1
        assert len(customer_b.facilities[0].units) == 3
        
        facility_b = customer_b.facilities[0]
        assert facility_b.city == 'Manchester'
        assert facility_b.country == 'England'
        
        # Check specific units
        unit_names = [unit.name for unit in facility_b.units]
        expected_names = ['Deep Freeze 1', 'Chilled Room 1', 'Chilled Room 2']
        assert set(unit_names) == set(expected_names)
    
    def test_export_to_yaml(self):
        """Test exporting customers to YAML"""
        customers = self.generator.generate_multiple_customers(3)
        
        with tempfile.NamedTemporaryFile(mode='w', suffix='.yaml', delete=False) as f:
            temp_path = f.name
        
        try:
            self.generator.export_to_yaml(temp_path)
            
            # Verify file was created and has content
            path = Path(temp_path)
            assert path.exists()
            assert path.stat().st_size > 0
            
            # Read file content as text first to check structure
            with open(temp_path, 'r') as f:
                content = f.read()
            
            # Basic structure checks
            assert 'customers:' in content
            assert len(customers) > 0
            
            # Try to load with unsafe loader to handle custom objects (for testing only)
            import yaml
            try:
                with open(temp_path, 'r') as f:
                    data = yaml.unsafe_load(f)  # Using unsafe_load for test
                
                assert 'customers' in data
                assert len(data['customers']) == 3
            except yaml.constructor.ConstructorError:
                # If YAML has serialization issues, just check file structure
                lines = content.split('\n')
                customer_lines = [line for line in lines if 'name:' in line]
                assert len(customer_lines) >= 3  # Should have at least 3 customer names
            
        finally:
            Path(temp_path).unlink()
    
    def test_get_customer(self):
        """Test retrieving generated customers"""
        customer = self.generator.generate_customer('RETRIEVE_TEST', 'pharmaceutical')
        
        retrieved = self.generator.get_customer('RETRIEVE_TEST')
        assert retrieved is not None
        assert retrieved.id == 'RETRIEVE_TEST'
        
        not_found = self.generator.get_customer('NOT_EXISTS')
        assert not_found is None
    
    def test_list_customers(self):
        """Test listing generated customer IDs"""
        initial_count = len(self.generator.list_customers())
        
        self.generator.generate_customer('LIST_TEST_1', 'pharmaceutical')
        self.generator.generate_customer('LIST_TEST_2', 'food_storage')
        
        customer_ids = self.generator.list_customers()
        assert len(customer_ids) == initial_count + 2
        assert 'LIST_TEST_1' in customer_ids
        assert 'LIST_TEST_2' in customer_ids

if __name__ == '__main__':
    pytest.main([__file__])