simulation/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
click>=8.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-dateutil>=2.8.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0